requirements.txt
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
openai==1.14.0
python-dotenv==1.2.2
python-multipart==0.0.22
uvicorn==0.42.0
